+++ servant-client/CHANGELOG.md
@@ -1,4 +1,4 @@
-HEAD
+0.5
 ----
 
 * Use the `text` package instead of `String`.
@@ -6,8 +6,11 @@ HEAD
 * Added support for `path` on `BaseUrl`.
 * `client` now takes an explicit `Manager` argument.
 * Use `http-api-data` instead of `Servant.Common.Text`
-* Client functions now consider any 2xx succesful.
+* Client functions now consider any 2xx successful.
 * Remove matrix params.
+* Added support for Basic authentication
+* Add generalized authentication support via the `AuthClientData` type family and
+  `AuthenticateReq` data type
 
 0.4.1
 -----