+++ servant/CHANGELOG.md
@@ -1,6 +1,7 @@
-HEAD
+0.5
 ----
 
+* Add `WithNamedConfig` combinator.
 * Add `HttpVersion`, `IsSecure`, `RemoteHost` and `Vault` combinators
 * Fix safeLink, so Header is not in fact required.
 * Add more instances for (:<|>)
@@ -9,6 +10,8 @@ HEAD
 * Add PlainText String MimeRender and MimeUnrender instances.
 * Add new `Verbs` combinator, and make all existing and new verb combinators
 type synonyms of it.
+* Add `BasicAuth` combinator to support Basic authentication
+* Add generalized authentication support
 
 0.4.2
 -----